Product Admin - Securing the views

+++ orderonline/views.py
@@ -4,6 +4,7 @@ from .models import MenuCategory, MenuItem, MenuItemIngredient, IngredientOption
 from .forms import AddToCartForm, ItemForm
 from django.forms import formset_factory
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import F
 
 def order_online(request):
@@ -48,13 +49,14 @@ def item_detail(request, item_id):
     })
 
 
-def product_management(request):
-    """ A view to return the product management page """
-    return render(request, 'orderonline/product_management.html')
 
-
+@login_required
 def add_item(request):
     """ Add a new item to the menu """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('order_online'))
+
     if request.method == 'POST':
         form = ItemForm(request.POST, request.FILES)
         if form.is_valid():
@@ -71,8 +73,13 @@ def add_item(request):
     }
     return render(request, template, context)
 
+@login_required
 def edit_item(request, item_id):
     """ Edit a product in the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('order_online'))
+
     item = get_object_or_404(MenuItem, pk=item_id)
     if request.method == 'POST':
         form = ItemForm(request.POST, request.FILES, instance=item)
@@ -94,8 +101,13 @@ def edit_item(request, item_id):
 
     return render(request, template, context)
 
+@login_required
 def delete_item(request, item_id):
     """ Delete a product from the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('order_online'))
+
     item = get_object_or_404(MenuItem, pk=item_id)
     item.delete()
     messages.success(request, 'Item deleted!')
